agents/seller-simple-agent/pyproject.toml
[tool.poetry]
name = "seller-simple-agent"
version = "0.1.0"
description = "Data selling agent with x402 payment-protected tools"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
strands-agents = {version = ">=1.0.0", extras = ["openai", "anthropic", "a2a"]}
payments-py = {version = ">=1.3.3", extras = ["strands", "langchain"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
openai = "^1.40.0"
python-dotenv = "^1.0.0"
langchain-openai = ">=0.3.0"
langgraph = ">=0.3.0"
anthropic = "^0.84.0"
groq = "^1.0.0"
python-multipart = "^0.0.22"
python-telegram-bot = "^21.0"

[tool.poetry.extras]
agentcore = ["bedrock-agentcore"]

[tool.poetry.scripts]
agent = "src.agent:main"
agent-agentcore = "src.agent_agentcore:main"
client = "src.client:main"
demo = "src.demo:main"
agent-a2a = "src.agent_a2a:main"
agent-a2a-agentcore = "src.agent_a2a_agentcore:main"
agent-langgraph = "src.server_langgraph:main"
mcp-server = "src.mcp_server:main"
web-demo = "src.web_demo:main"
telegram-bot = "src.telegram_bot:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
